Reverts Model and TypeTraits. Moving forward with UCB.
