--- include/AIToolbox/MDP/TypeTraitsEigen.hpp
@@ -1,87 +1,0 @@
-#ifndef AI_TOOLBOX_MDP_TYPE_TRAITS_EIGEN_HEADER_FILE
-#define AI_TOOLBOX_MDP_TYPE_TRAITS_EIGEN_HEADER_FILE
-
-// #include <AIToolbox/MDP/Types.hpp>
-#include <AIToolbox/MDP/TypeTraits.hpp>
-
-namespace AIToolbox::MDP {
-    /**
-     * @brief This struct represents the required interface that allows MDP algorithms to leverage Eigen.
-     *
-     * This struct is used to check interfaces of classes in templates.
-     * In particular, this struct tests for the interface of an MDP model
-     * which uses Eigen matrices internally.
-     * The interface must be implemented and be public in the parameter
-     * class. The interface is the following:
-     *
-     * - T getTransitionFunction(size_t a) const : Returns the transition function for a given action as a matrix SxS', where T is some Eigen matrix type.
-     * - R getRewardFunction(size_t a) const : Returns the reward function for a given action as a matrix SxS', where R is some Eigen matrix type.
-     *
-     * In addition the MDP needs to respect the interface for the MDP model.
-     *
-     * \sa MDP::is_model
-     *
-     * is_model_eigen<M>::value will be equal to true is M implements the interface,
-     * and false otherwise.
-     *
-     * @tparam M The class to test for the interface.
-     */
-    template <typename M>
-    struct is_model_eigen {
-        private:
-            // With this macro we can find out the return type of a given member function; we use it
-            // so that we can check whether the class offers methods which return Eigen types, so we
-            // can enable the high-performance algorithm variants.
-            #define RETVAL_EXTRACTOR(fun_name)                                                                                      \
-                                                                                                                                    \
-            template <typename Z, typename ...Args> static auto fun_name##RetType(Z* z) ->                                          \
-                                                                remove_cv_ref_t<decltype(z->fun_name(std::declval<Args>()...))>;    \
-                                                                                                                                    \
-            template <typename Z, typename ...Args> static auto fun_name##RetType(...) -> int
-
-            RETVAL_EXTRACTOR(getTransitionFunction);
-            RETVAL_EXTRACTOR(getRewardFunction);
-
-            // The template parameters here must match the ones used in the test function!
-            // So const M if the function is const, and then the parameter types.
-            using F = decltype(getTransitionFunctionRetType<const M, size_t>(0));
-            using R = decltype(getRewardFunctionRetType<const M>(0));
-
-            template <typename Z> static constexpr auto test(int) -> decltype(
-
-                    static_cast<const F & (Z::*)(size_t) const>         (&Z::getTransitionFunction),
-                    static_cast<const R & (Z::*)()       const>         (&Z::getRewardFunction),
-
-                    bool()
-            ) { return true; }
-
-            template <typename Z> static constexpr auto test(...) -> bool
-            { return false; }
-
-            #undef RETVAL_EXTRACTOR
-
-        public:
-            enum {
-                value = is_model_v<M> && test<M>(0) &&
-                        std::is_base_of_v<Eigen::EigenBase<F>, F> &&
-                        std::is_base_of_v<Eigen::EigenBase<R>, R>
-            };
-    };
-    template <typename M>
-    inline constexpr bool is_model_eigen_v = is_model_eigen<M>::value;
-
-    /**
-     * @brief This struct verifies that a class satisfies the is_model interface but not the is_model_eigen interface.
-     *
-     * @tparam M The class to test for the interface.
-     */
-    template <typename M>
-    struct is_model_not_eigen {
-        public:
-            enum { value = is_model_v<M> && !is_model_eigen_v<M> };
-    };
-    template <typename M>
-    inline constexpr bool is_model_not_eigen_v = is_model_not_eigen<M>::value;
-}
-
-#endif
